refactor: 베팅 금액 패키지 이동
